Migrate checker to MiscDiagnostics
